Fixed bugs for RegisterPage - changed to use RedwoodJS components for cleaner code and client-side error handling
