+++ web/src/pages/RegisterPage/RegisterPage.css
@@ -1,0 +1,95 @@
+p{
+  font-family: sans-serif;
+}
+.title{
+  font-size: 40px;
+  font-weight: bold;
+}
+.home{
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  margin-top: 20px;
+  background-color: paleturquoise;
+  height: 1080px;
+}
+
+.field{
+  padding: 10px;
+  margin: 5px;
+  border-radius: 15px;
+  border: 1px solid #ccc; /* Regular border color */
+  font-size: 16px; /* Consistent font size */
+  transition: border-color 0.3s ease; /* Smooth transition for border color */
+}
+
+.error-field{
+  padding: 10px;
+  margin: 5px;
+  border-radius: 15px;
+  border: 1px solid red; /* Regular border color */
+  font-size: 16px; /* Consistent font size */
+  transition: border-color 0.3s ease; /* Smooth transition for border color */
+}
+
+.input-wrapper.error {
+  border-color: red;
+}
+
+.input-wrapper.error .field:focus {
+  border-color: darkred;
+}
+
+.error-text{
+  color: red;
+  font-size: 12px;
+  margin-top: 5px;
+}
+
+.subtext{
+  font-size: 30px;
+}
+
+
+.signup{
+  margin-top: 20px;
+}
+
+
+/* Button Styling */
+.button {
+  background: linear-gradient(145deg, #6e7dff, #5560ea); /* Gradient background */
+  color: white;                          /* Text color */
+  padding: 12px 25px;                    /* Padding around the text */
+  border: none;                          /* No border */
+  border-radius: 30px;                   /* Rounded corners */
+  font-size: 16px;                       /* Font size */
+  font-weight: 600;                      /* Bold font */
+  cursor: pointer;                      /* Pointer cursor on hover */
+  box-shadow: 2px 4px 10px rgba(0, 0, 0, 0.1); /* Shadow for depth */
+  transition: all 0.3s ease;             /* Smooth transition for hover effects */
+  margin: auto
+}
+
+
+.button:hover {
+  background: linear-gradient(145deg, #5560ea, #6e7dff); /* Reversed gradient for hover */
+  box-shadow: 4px 8px 20px rgba(0, 0, 0, 0.2); /* Bigger shadow on hover */
+  transform: translateY(-3px);              /* Slightly lifts the button */
+}
+
+
+.button:active {
+  transform: translateY(1px); /* Button sinks down slightly when clicked */
+  box-shadow: 2px 4px 10px rgba(0, 0, 0, 0.1); /* Reset shadow */
+}
+
+.separator{
+  height: 20px;
+}
+
+.signup{
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
