--- web/src/pages/SignUpPage/SignUpPage.stories.tsx
@@ -1,13 +1,0 @@
-import type { Meta, StoryObj } from '@storybook/react'
-
-import SignUpPage from './SignUpPage'
-
-const meta: Meta<typeof SignUpPage> = {
-  component: SignUpPage,
-}
-
-export default meta
-
-type Story = StoryObj<typeof SignUpPage>
-
-export const Primary: Story = {}
